Placement + Data
Placement still needs a lot of work. It's a prototype at this point and
would be executed after the ui is created for constructing buildings.
For some reason buildings are being added to the bases arrayList even
though the coordinates should not contain any building. I think it has
something to do with the mouse position in  the raycast mouse script.